+++ src/aws/output.tf
@@ -9,3 +9,7 @@ output "iam_access_key" {
 output "iam_access_secret" {
   value = "${aws_iam_access_key.funmooc_access_key.secret}"
 }
+
+output "acm_certificate_validation" {
+  value = "${aws_acm_certificate.certificate.domain_validation_options}"
+}